pydanticai-101/pyproject.toml
[project]
name = "pydanticai-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[logfire]>=0.0.13",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = []
